fix(calendar): fixed few bugs and refactored some codes
