+++ src/utils/useEvent.ts
@@ -1,7 +1,8 @@
 import { useContext } from "react"
-import { EventContext } from "./Events"
+import { EventContext } from "../context/Events"
 
 export const EVENT_COLORS = ["red", "green", "blue"] as const
+
 export function useEvent() {
   const value = useContext(EventContext)
